Add support for assignments context.

+++ packages/rbac-http/src/adapters/RbacHttpAssignmentAdapter.ts
@@ -1,6 +1,6 @@
-import {AxiosInstance} from 'axios';
+import { AxiosInstance } from 'axios';
 
-import { RbacAssignment, RbacAssignmentAdapter, RbacItem, RbacUserId } from '@brainstaff/rbac';
+import { defaultRbacAssignmentContextId, RbacAssignment, RbacAssignmentAdapter, RbacAssignmnentContextId, RbacItem, RbacUserId } from '@brainstaff/rbac';
 
 import { rethrow } from '../utils/rethrow';
 
@@ -14,40 +14,42 @@ export default class RbacHttpAssignmentAdapter implements RbacAssignmentAdapter 
   }
 
   @rethrow
-  async store(values: RbacAssignment[]) {
-    await this.client.post(`/rbac/assignments`, { rbacAssignments: values });
+  async store(raw: RbacAssignment[]) {
+    const all = raw.map(x => new RbacAssignment(x));
+    await this.client.post(`/rbac/assignments`, { rbacAssignments: all });
   }
 
   @rethrow
-  async load() {
-    const res = await this.client.get(`/rbac/assignments`);
+  async load(contextId = defaultRbacAssignmentContextId) {
+    const res = await this.client.get(`/rbac/assignments`, { params: { contextId } });
     return res.data.map((x: any) => new RbacAssignment(x));
   }
 
   @rethrow
-  async create(userId: RbacUserId, role: RbacItem['name']) {
-    await this.client.post(`/rbac/assignments`, { userId, role });
+  async create(raw: RbacAssignment) {
+    const one = new RbacAssignment(raw);
+    await this.client.post(`/rbac/assignments`, one);
   }
 
   @rethrow
-  async find(userId: RbacUserId, role: RbacItem['name']) {
-    const res = await this.client.get(`/rbac/assignments/${userId}/${role}`);
+  async find(userId: RbacUserId, role: RbacItem['name'], contextId = defaultRbacAssignmentContextId) {
+    const res = await this.client.get(`/rbac/assignments/${userId}/${role}`, { params: { contextId } });
     return res.data == null ? null : new RbacAssignment(res.data);
   }
 
   @rethrow
-  async findByUserId(userId: RbacUserId) {
-    const res = await this.client.get(`/rbac/assignments/${userId}`);
+  async findByUserId(userId: RbacUserId, contextId = defaultRbacAssignmentContextId) {
+    const res = await this.client.get(`/rbac/assignments/${userId}`, { params: { contextId } });
     return res.data.map((x: any) => new RbacAssignment(x));
   }
 
   @rethrow
-  async delete(userId: RbacUserId, role: RbacItem['name']) {
-    await this.client.delete(`/rbac/assignments/${userId}/${role}`);
+  async delete(userId: RbacUserId, role: RbacItem['name'], contextId = defaultRbacAssignmentContextId) {
+    await this.client.delete(`/rbac/assignments/${userId}/${role}`, { params: { contextId } });
   }
 
   @rethrow
-  async deleteByUser(userId: RbacUserId) {
-    await this.client.delete(`/rbac/assignments/${userId}`);
+  async deleteByUser(userId: RbacUserId, contextId = defaultRbacAssignmentContextId) {
+    await this.client.delete(`/rbac/assignments/${userId}`, { params: { contextId } });
   }
 }
